--- blockchain/browsersync/daoevent/ScanDaoEventLogOnPolygon.go
@@ -1,92 +1,0 @@
-package daoevent
-
-import (
-	"math/big"
-	"payment-bridge/common/constants"
-	"payment-bridge/common/utils"
-	"payment-bridge/database"
-	"payment-bridge/logs"
-	models2 "payment-bridge/models"
-	"strconv"
-	"sync"
-	"time"
-)
-
-func ScanEventFromChainAndSaveDataToDbForPolygon() {
-	startScanBlockNo := utils.GetStartBlockNo(constants.TABLE_NAME_EVENT_POLYGON, GetConfig().PolygonMainnetNode.StartFromBlockNo)
-
-	for {
-		var mutex sync.Mutex
-		mutex.Lock()
-		var blockNoCurrent *big.Int
-		var err error
-		var getBlockFlag bool = true
-		for getBlockFlag {
-			blockNoCurrent, err = WebConn.GetBlockNumber()
-			if err != nil {
-				ClientInit()
-				logs.GetLogger().Error(err)
-				time.Sleep(5 * time.Second)
-				continue
-			}
-			if err == nil {
-				getBlockFlag = false
-			}
-		}
-
-		blockScanRecord := new(models2.BlockScanRecord)
-		whereCondition := "network_type='" + constants.NETWORK_TYPE_POLYGON + "'"
-		blockScanRecordList, err := blockScanRecord.FindLastCurrentBlockNumber(whereCondition)
-		if err != nil {
-			logs.GetLogger().Error(err)
-			startScanBlockNo = GetConfig().PolygonMainnetNode.StartFromBlockNo
-		}
-		if len(blockScanRecordList) > 0 {
-			if blockScanRecordList[0].LastCurrentBlockNumber <= blockNoCurrent.Int64() {
-				startScanBlockNo = blockScanRecordList[0].LastCurrentBlockNumber
-			} else {
-				startScanBlockNo = GetConfig().PolygonMainnetNode.StartFromBlockNo
-			}
-			blockScanRecord.ID = blockScanRecordList[0].ID
-		}
-
-		for {
-			start := startScanBlockNo
-			end := start + GetConfig().PolygonMainnetNode.ScanStep
-			if startScanBlockNo > blockNoCurrent.Int64() {
-				break
-			}
-			err = ScanPolygonEventFromChainAndSaveEventLogData(start, end)
-			if err != nil {
-				logs.GetLogger().Error(err)
-				continue
-			}
-
-			if end >= blockNoCurrent.Int64() {
-				blockScanRecord.LastCurrentBlockNumber = blockNoCurrent.Int64()
-			} else {
-				blockScanRecord.LastCurrentBlockNumber = end
-			}
-
-			blockScanRecord.NetworkType = constants.NETWORK_TYPE_POLYGON
-			blockScanRecord.UpdateAt = strconv.FormatInt(utils.GetEpochInMillis(), 10)
-
-			err = database.SaveOne(blockScanRecord)
-			if err != nil {
-				logs.GetLogger().Error(err)
-				continue
-			}
-			start = end
-			startScanBlockNo = end
-			if end >= blockNoCurrent.Int64() {
-				break
-			}
-		}
-
-		getBlockFlag = true
-		mutex.Unlock()
-
-		time.Sleep(time.Second * GetConfig().PolygonMainnetNode.CycleTimeInterval)
-		logs.GetLogger().Info("-------------------------polygon----------------------------")
-	}
-}
