--- blockchain/browsersync/daoevent/polygon_registry.go
@@ -1,17 +1,0 @@
-package daoevent
-
-import (
-	"payment-bridge/models"
-)
-
-func Init() {
-	models.Register(&models.RegInfo{
-		Network:                           "Polygon",
-		CoinName:                          "MATIC",
-		ScanEventFromChainAndSaveDataToDb: ScanEventFromChainAndSaveDataToDb,
-	})
-}
-
-func ScanEventFromChainAndSaveDataToDb() {
-	ScanEventFromChainAndSaveDataToDbForPolygon()
-}
