+++ blockchain/browsersync/scanlockpayment/bsc/bsc_setup.go
