+++ blockchain/browsersync/scanlockpayment/polygon/PolygonDaoEventLogService.go
@@ -1,4 +1,4 @@
-package nbai
+package polygon
 
 import (
 	"context"
@@ -7,37 +7,30 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"math/big"
+	"payment-bridge/blockchain/browsersync/scanlockpayment/goerli"
 	"payment-bridge/common/utils"
 	"payment-bridge/database"
 	"payment-bridge/logs"
 	"payment-bridge/models"
-	"payment-bridge/on-chain/goBind"
 	"strconv"
 	"strings"
 	"time"
 )
 
-/**
- * created on 08/20/21.
- * author: nebula-ai-zhiqiang
- * Copyright defined in payment-bridge/LICENSE
- */
-
 // EventLogSave Find the event that executed the contract and save to db
-func ScanNbaiEventFromChainAndSaveEventLogData(blockNoFrom, blockNoTo int64) error {
+func ScanDaoEventFromChainAndSaveEventLogData(blockNoFrom, blockNoTo int64) error {
 	//read contract api json file
-	logs.GetLogger().Println("nbai blockNoFrom=" + strconv.FormatInt(blockNoFrom, 10) + "--------------blockNoTo=" + strconv.FormatInt(blockNoTo, 10))
-	//paymentAbiString, err := utils.ReadContractAbiJsonFile(goBind.StateSenderABI)
-	paymentAbiString, err := abi.JSON(strings.NewReader(string(goBind.StateSenderABI)))
+	logs.GetLogger().Println("scan dao event on polygon : blockNoFrom=" + strconv.FormatInt(blockNoFrom, 10) + "--------------blockNoTo=" + strconv.FormatInt(blockNoTo, 10))
+	daoEventAbiString, err := utils.ReadContractAbiJsonFile(goerli.SwanPaymentAbiJson)
 	if err != nil {
 		logs.GetLogger().Error(err)
 		return err
 	}
 
 	//SwanPayment contract address
-	contractAddress := common.HexToAddress(GetConfig().NbaiMainnetNode.PaymentContractAddress)
+	contractAddress := common.HexToAddress(GetConfig().PolygonMainnetNode.PaymentContractAddress)
 	//SwanPayment contract function signature
-	contractFunctionSignature := GetConfig().NbaiMainnetNode.ContractFunctionSignature
+	contractFunctionSignature := GetConfig().PolygonMainnetNode.ContractFunctionSignature
 
 	//test block no. is : 5297224
 	query := ethereum.FilterQuery{
@@ -63,40 +56,40 @@ func ScanNbaiEventFromChainAndSaveEventLogData(blockNoFrom, blockNoTo int64) err
 		}
 	}
 
+	contractAbi, err := abi.JSON(strings.NewReader(daoEventAbiString))
+	if err != nil {
+		logs.GetLogger().Error(err)
+		return err
+	}
+
 	for _, vLog := range logsInChain {
 		//if log have this contractor function signer
 		if vLog.Topics[0].Hex() == contractFunctionSignature {
-			eventList, err := models.FindEventNbai(&models.EventNbai{TxHash: vLog.TxHash.Hex(), BlockNo: vLog.BlockNumber}, "id desc", "10", "0")
+			eventList, err := models.FindDaoEventLog(&models.DaoEventLog{TxHash: vLog.TxHash.Hex(), BlockNo: vLog.BlockNumber}, "id desc", "10", "0")
 			if err != nil {
 				logs.GetLogger().Error(err)
 				continue
 			}
 			if len(eventList) <= 0 {
-				receiveMap := map[string]interface{}{}
-				err = paymentAbiString.UnpackIntoMap(receiveMap, "StateSynced", vLog.Data)
+				var event = new(models.DaoEventLog)
+				dataList, err := contractAbi.Unpack("LockPayment", vLog.Data)
 				if err != nil {
 					logs.GetLogger().Error(err)
-					continue
 				}
-				var event = new(models.EventNbai)
-				addrInfo, err := utils.GetFromAndToAddressByTxHash(WebConn.ConnWeb, big.NewInt(GetConfig().NbaiMainnetNode.ChainID), vLog.TxHash)
+
+				addrInfo, err := utils.GetFromAndToAddressByTxHash(WebConn.ConnWeb, big.NewInt(GetConfig().PolygonMainnetNode.ChainID), vLog.TxHash)
 				if err != nil {
 					logs.GetLogger().Error(err)
 				} else {
-					event.AddressFrom = addrInfo.AddrFrom
-					event.AddressTo = addrInfo.AddrTo
+					event.DaoAddress = addrInfo.AddrFrom
 				}
-
 				event.BlockNo = vLog.BlockNumber
 				event.TxHash = vLog.TxHash.Hex()
-				event.ContractName = "SwanPayment"
-				event.ContractAddress = contractAddress.String()
-				event.BytesData = receiveMap["data"].([]byte)
-				event.CreateAt = strconv.FormatInt(utils.GetEpochInMillis(), 10)
+				event.PayloadCid = dataList[0].(string)
+
 				err = database.SaveOneWithTransaction(event)
 				if err != nil {
 					logs.GetLogger().Error(err)
-					continue
 				}
 			}
 		}
